Fix #2888: error when commit/rollback on invalid transaction; fix error msg on nested transaction (#3469)
* throw exceptions for nested transactions; correct error message
* add todo message for error improving bug in shell
* remove restriction on manual transaction on copy statement; remove incorrect rollback in execute interface; py lint; clean java test

+++ src/storage/wal_replayer.cpp
@@ -224,7 +224,10 @@ void WALReplayer::replayDropCatalogEntryRecord(const WALRecord& walRecord) {
     case TableType::REL:
     case TableType::RDF: {
         clientContext.getCatalog()->dropTableSchema(&DUMMY_WRITE_TRANSACTION, tableID);
-        clientContext.getStorageManager()->dropTable(tableID);
+        // During recovery, storageManager does not exist.
+        if (clientContext.getStorageManager()) {
+            clientContext.getStorageManager()->dropTable(tableID);
+        }
     } break;
     default: {
         KU_UNREACHABLE;
